+++ lib/src/injections/nano_default_injections.dart
@@ -1,5 +1,6 @@
 import 'package:get_it/get_it.dart';
 import '../http/nano_http_client.dart';
+import '../pagination/nano_pagination.dart';
 import 'nano_injections.dart';
 
 /// Convenient type alias for [NanoDefaultInjections].
@@ -7,21 +8,25 @@ typedef NanoCoreInjections = NanoDefaultInjections;
 
 /// Default dependency injection container for framework-level services.
 ///
-/// Registers essential framework singletons such as [NanoHttpClient]
-/// into [GetIt].
+/// Registers essential framework singletons such as [NanoHttpClient] and
+/// global [NanoPagination] strategy into [GetIt].
 class NanoDefaultInjections extends NanoInjections {
   /// Creates a [NanoDefaultInjections] scope.
   const NanoDefaultInjections({
     this.client,
+    this.pagination,
     super.scope = 'nano_default_global',
   });
 
   /// The global [NanoHttpClient] instance to be registered.
   final NanoHttpClient? client;
 
+  /// The optional default global [NanoPagination] strategy.
+  final NanoPagination? pagination;
+
   @override
   void binds(GetIt i) {
-    init(i, client: client);
+    init(i, client: client, pagination: pagination);
   }
 
   /// Initializes default framework dependencies directly inside an existing
@@ -29,9 +34,13 @@ class NanoDefaultInjections extends NanoInjections {
   static void init(
     GetIt i, {
     NanoHttpClient? client,
+    NanoPagination? pagination,
   }) {
     if (client != null && !i.isRegistered<NanoHttpClient>()) {
       i.registerLazySingleton<NanoHttpClient>(() => client);
+    }
+    if (pagination != null && !i.isRegistered<NanoPagination>()) {
+      i.registerLazySingleton<NanoPagination>(() => pagination);
     }
   }
 
@@ -39,7 +48,8 @@ class NanoDefaultInjections extends NanoInjections {
   /// globally at app startup (e.g. in `main()`).
   static void register({
     NanoHttpClient? client,
+    NanoPagination? pagination,
   }) {
-    init(GetIt.I, client: client);
+    init(GetIt.I, client: client, pagination: pagination);
   }
 }